chore(useuniqueid): convert tests to RTL
